src/strategies/llm_strategies/gemini.py
"""
Gemini-based trading strategy using Google's Gemini models.

This module implements a trading strategy that uses Google's Gemini models
with advanced multimodal capabilities for comprehensive market analysis.
"""

import google.generativeai as genai
import time
import json
from typing import List, Dict, Any, Optional
import logging

from .base_llm import BaseLLMStrategy, LLMResponse
from ..base import RiskParameters

logger = logging.getLogger(__name__)


class GeminiStrategy(BaseLLMStrategy):
    """
    Trading strategy using Google's Gemini models.
    
    This strategy leverages Gemini's advanced reasoning and multimodal
    capabilities to analyze market data and generate informed trading
    recommendations with detailed reasoning.
    """
    
    def __init__(
        self,
        api_key: str,
        model: str = "gemini-1.5-pro",
        risk_params: RiskParameters = None,
        enable_thinking: bool = True,
        enable_research: bool = True,
        max_recommendations: int = 8
    ):
        """
        Initialize Gemini strategy.
        
        Args:
            api_key: Google AI API key
            model: Gemini model to use
            risk_params: Risk management parameters
            enable_thinking: Enable thinking mode
            enable_research: Enable research mode
            max_recommendations: Maximum stock recommendations
        """
        super().__init__(
            name=f"Gemini-{model.split('-')[-1]}",
            model_name=model,
            risk_params=risk_params,
            enable_thinking=enable_thinking,
            enable_research=enable_research,
            max_recommendations=max_recommendations
        )
        
        genai.configure(api_key=api_key)
        self.model = genai.GenerativeModel(model)
        self.model_name_full = model
    
    def _call_llm(
        self, 
        prompt: str, 
        system_message: Optional[str] = None,
        temperature: float = 0.1
    ) -> LLMResponse:
        """
        Make API call to Google Gemini.
        
        Args:
            prompt: The main prompt for Gemini
            system_message: System instruction for the model
            temperature: Sampling temperature
            
        Returns:
            LLMResponse with Gemini's output
            
        Raises:
            Exception: If API call fails
        """
        pass
    
    def _create_system_message(self) -> str:
        """
        Create system instruction optimized for Gemini's capabilities.
        
        Returns:
            System instruction that leverages Gemini's strengths
        """
        pass
    
    def _create_comprehensive_prompt(
        self,
        universe: List[str],
        market_context: str,
        portfolio_context: str
    ) -> str:
        """
        Create a comprehensive analysis prompt for Gemini.
        
        Args:
            universe: Available stock tickers
            market_context: Market analysis context
            portfolio_context: Current portfolio context
            
        Returns:
            Detailed prompt for multimodal market analysis
        """
        pass
    
    def _configure_generation_settings(self, temperature: float = 0.1) -> Dict[str, Any]:
        """
        Configure Gemini generation settings.
        
        Args:
            temperature: Sampling temperature
            
        Returns:
            Generation configuration dictionary
        """
        pass
    
    def _handle_gemini_safety_settings(self) -> List[Dict[str, Any]]:
        """
        Configure safety settings for financial analysis.
        
        Returns:
            List of safety setting configurations
        """
        pass
    
    def _parse_gemini_response(self, response) -> LLMResponse:
        """
        Parse Gemini's response format.
        
        Args:
            response: Raw Gemini response object
            
        Returns:
            Structured LLMResponse object
        """
        pass
    
    def _extract_structured_output(self, content: str) -> Dict[str, Any]:
        """
        Extract structured output from Gemini's response.
        
        Args:
            content: Raw response content
            
        Returns:
            Parsed structured output
        """
        pass